Revert "Extensibility: Make Block Bindings work with `editor.BlockEdit` hook …"
This reverts commit c517e410017f4d65a7c6f03a31c5c2fa15cbbd65.
