+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -29,7 +29,7 @@ import { useBlockRefProvider } from './use-block-refs';
 import { useIntersectionObserver } from './use-intersection-observer';
 import { useScrollIntoView } from './use-scroll-into-view';
 import { useFlashEditableBlocks } from '../../use-flash-editable-blocks';
-import { canBindBlock } from '../../../utils/block-bindings';
+import { canBindBlock } from '../../../hooks/use-bindings-attributes';
 import { useFirefoxDraggableCompatibility } from './use-firefox-draggable-compatibility';
 
 /**
